chore: rename folder to trigger checks
