+++ lib/components/SelectionControl/checkboxs/checkbox-single/checkbox.scss
