+++ lib/components/SelectionControl/index.ts
@@ -1,2 +1,2 @@
-export * from "./checkbox/checkbox-single";
-export * from "./checkbox/checkbox-group";
+export * from "./checkboxs/checkbox-single";
+export * from "./checkboxs/checkbox-group";
